making install_ruby and install_compass rerunnable.

+++ centos/install_compass.sh
@@ -1,2 +1,5 @@
-
-gem install compass
+if [ -e /usr/bin/compass ]; then
+    echo "compass already installed... skipping."
+else
+    gem install compass
+fi
